spark-model: split the DFlash loader tests out for the 500-LoC cap
The DFlash2 fail-closed guard and its test pushed dflash_loader.rs to 560 LoC
over the CI cap. Move the test module to dflash_loader/tests.rs.
